+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # master 
 
-* functions filter: index and rindex
+* functions filter: index, rindex, and round
 * functions filter: obj.to_json => JSON.stringify(obj)
 * support numbered parameters (numblocks)
 * Array.new(size, default) => new Array(size).fill(default)